pyproject.toml
[project]
name = "mikros-server"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=3.2.3",
    "jsonschema>=4.23.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mikros_server*"]

[tool.setuptools.package-data]
mikros_server = ["workflows/*.yaml", "*.md"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
